feat: Principios de autorização e autenticação
Acesso a registros agora apenas ocorrem para usuários logados
e que sejam proprietários dos mesmos, impedindo acessos indevidos.
A implementação atual é simplificada, porém utiliza de recursos
recomendados do framework, evitando repetição de código.

+++ app/Http/Controllers/Registro/FixoController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Registro;
 
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
 use App\Models\Categorizadores\Gerais\Categoria;
 use App\Models\Categorizadores\Gerais\Localizacao;
 use App\Models\Categorizadores\Gerais\Nivel_imp;
@@ -13,14 +14,16 @@ use App\Models\Personas\Realizador;
 use App\Models\Recursos\RegistroFixo;
 use BcMath\Number;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class FixoController extends Controller
 {
     //Métodos de recurso
     public function index()
     {
-        $registros = RegistroFixo::where("cd_usuario", 2)
-            ->orderBy("nm_registroFixo")
+        $registros = RegistroFixo::where("cd_usuario", Auth::user()->cd_usuario)
+            ->orderBy("cd_nivel_imp", "desc")
+            ->orderBy("nm_registroFixo", "asc")
             ->paginate(9);
         return view("registro.fixo.index", [
             "registros" => $registros,
@@ -28,6 +31,9 @@ class FixoController extends Controller
     }
     public function show(RegistroFixo $registroFixo)
     {
+        //Autorizando accesso ao recurso ou não
+        Gate::authorize("access-registroFixo", $registroFixo);
+
         return view("registro.fixo.show", [
             "registro" => $registroFixo,
         ]);
@@ -52,10 +58,12 @@ class FixoController extends Controller
     }
     public function edit(RegistroFixo $registroFixo)
     {
+        Gate::authorize('access-registroFixo',$registroFixo);
         $metodosProprios = $registroFixo
             ->metodo_pagamento()
             ->pluck("metodo_pagamento.cd_tipo_metodo")
             ->toArray();
+
         return view("registro.fixo.edit", [
             "registro" => $registroFixo,
             "metodosProprios" => $metodosProprios,
@@ -68,13 +76,14 @@ class FixoController extends Controller
             "realizadores" => Realizador::all(),
         ]);
     }
-    public function update(Request $request, int $id)
+    public function update(Request $request, RegistroFixo $registroFixo)
     {
-        //To-do
+        Gate::authorize('access-registroFixo',$registroFixo);
         return dd($request->all());
     }
     public function destroy(RegistroFixo $registroFixo)
     {
+        Gate::authorize('access-registroFixo',$registroFixo);
         $registroFixo->delete();
         return redirect(route("registroFixo.index"));
     }
